requirements.txt
fastapi
uvicorn[standard]
requests
feedparser
python-dotenv
sentence-transformers
numpy
